requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imap-tools==1.11.0
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pandas-stubs==2.3.3.251219
psycopg2==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
six==1.17.0
types-pytz==2025.2.0.20251108
tzdata==2025.2
urllib3==2.6.3
xlrd==2.0.2

maxapi